pyproject.toml
[tool.poetry]
name = "timetable"
version = "0.1.0"
description = ""
authors = ["Lawrence Ephrim <ephrimlawrence@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
django-widget-tweaks = "^1.4"
itsdangerous = "^1.1"
django-datatable-view = "^0.9.0"
django-tables2 = "^2.1"
tablib = "^0.14.0"
django-filter = "^2.2"
django-bootstrap4 = "^1.0"
django-safedelete = "^0.5.2"
arrow = "^0.15.2"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
isort = "^4.3"
flake8 = "^3.7"
mypy = "^0.730.0"
pytest-cov = "^2.7"
pre-commit = "^1.18"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
bpython = "^0.18.0"


[tool.black]
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | timetable/timetable/administrator/migrations
    | timetable/timetable/staff/migrations
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

